LayoutTests/http/tests/cache/disk-cache/disk-cache-request-headers.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cache-test.js"></script>
<body>
<script>

var testMatrix =
[
 [
  { responseHeaders: {'Cache-control': 'max-age=0' } },
  { responseHeaders: {'Cache-control': 'max-age=100' } },
  ],
 [
  {},
  { requestHeaders: {'Cache-control': 'no-cache' } },
  ],
 [
  {},
  { requestHeaders: {'Cache-control': 'no-store' } },
  ],
 [
  {},
  { requestHeaders: {'Cache-control': 'max-age=0' } },
  { requestHeaders: {'Cache-control': 'max-age=100' } },
  ],
 ];

description("Test permutations of various cache headers");

var tests = generateTests(testMatrix);

debug("running " + tests.length + " tests");
debug("");

runTests(tests);

</script>
<script src="/js-test-resources/js-test-post.js"></script>
