LayoutTests/http/tests/cache/disk-cache/disk-cache-request-headers-expected.txt

Test permutations of various cache headers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 24 tests

response headers: {"Cache-control":"max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"no-cache, no-store"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"no-cache, no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"no-cache, max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"no-cache, max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"no-store, max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"no-store, max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"no-cache, no-store, max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"no-cache, no-store, max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"max-age=100"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"max-age=100"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"no-cache, max-age=100"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"no-cache, max-age=100"}
response source: Network

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"no-store, max-age=100"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"no-store, max-age=100"}
response source: Network

response headers: {"Cache-control":"max-age=0"}
request headers: {"Cache-control":"no-cache, no-store, max-age=100"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
request headers: {"Cache-control":"no-cache, no-store, max-age=100"}
response source: Network

PASS successfullyParsed is true

TEST COMPLETE
